0981588 COMP: Ivars of wrong type.
ab4fadd BUG: If data has any polys, ignore all lines.

+++ Imaging/vtkImageStencilToImage.h
@@ -76,8 +76,8 @@ protected:
                           vtkInformationVector **,
                           vtkInformationVector *);
 
-  int OutsideValue;
-  int InsideValue;
+  double OutsideValue;
+  double InsideValue;
   int OutputScalarType;
 
   virtual int FillInputPortInformation(int, vtkInformation*);
